Moved SpectralSimilarity to util/scan/similarity (from maths)
