+++ src/main/java/net/sf/mzmine/util/scans/similarity/impl/composite/CompositeCosineSpectralSimilarity.java
@@ -16,7 +16,7 @@
  * USA
  */
 
-package net.sf.mzmine.util.maths.similarity.spectra.impl.composite;
+package net.sf.mzmine.util.scans.similarity.impl.composite;
 
 import java.util.Arrays;
 import java.util.List;
@@ -27,10 +27,10 @@ import net.sf.mzmine.datamodel.DataPoint;
 import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.parameters.parametertypes.tolerances.MZTolerance;
 import net.sf.mzmine.util.maths.similarity.Similarity;
-import net.sf.mzmine.util.maths.similarity.spectra.SpectralSimilarity;
-import net.sf.mzmine.util.maths.similarity.spectra.SpectralSimilarityFunction;
-import net.sf.mzmine.util.maths.similarity.spectra.Weights;
 import net.sf.mzmine.util.scans.ScanAlignment;
+import net.sf.mzmine.util.scans.similarity.SpectralSimilarity;
+import net.sf.mzmine.util.scans.similarity.SpectralSimilarityFunction;
+import net.sf.mzmine.util.scans.similarity.Weights;
 
 /**
  * Similar to NIST search algorithm for GC-MS data with lots of signals (more an identity check than
