+++ src/main/java/net/sf/mzmine/util/scans/similarity/impl/composite/CompositeCosineSpectralSimilarityParameters.java
@@ -16,14 +16,14 @@
  * USA
  */
 
-package net.sf.mzmine.util.maths.similarity.spectra.impl.composite;
+package net.sf.mzmine.util.scans.similarity.impl.composite;
 
 import java.text.DecimalFormat;
 import net.sf.mzmine.parameters.Parameter;
 import net.sf.mzmine.parameters.impl.SimpleParameterSet;
 import net.sf.mzmine.parameters.parametertypes.ComboParameter;
 import net.sf.mzmine.parameters.parametertypes.DoubleParameter;
-import net.sf.mzmine.util.maths.similarity.spectra.Weights;
+import net.sf.mzmine.util.scans.similarity.Weights;
 
 /**
  *
