+++ src/main/java/net/sf/mzmine/util/spectraldb/entry/SpectralDBPeakIdentity.java
@@ -27,7 +27,7 @@ import net.sf.mzmine.datamodel.impl.SimplePeakIdentity;
 import net.sf.mzmine.util.DataPointSorter;
 import net.sf.mzmine.util.SortingDirection;
 import net.sf.mzmine.util.SortingProperty;
-import net.sf.mzmine.util.maths.similarity.spectra.SpectralSimilarity;
+import net.sf.mzmine.util.scans.similarity.SpectralSimilarity;
 
 public class SpectralDBPeakIdentity extends SimplePeakIdentity {
   private static final DecimalFormat COS_FORM = new DecimalFormat("0.000");
